Chat display issue and survey completion bug fixes.
Resolution #1 - The chat toolbar no longer attaches to the top of the
screen and hides chat history after the image options dialog (camera
icon) is tapped and hidden (by sending a photo or tapping cancel).

Resolution #2 - In the demo app, forms will now complete successfully
and display the post-survey "thank you" screen.

+++ EXPERTconnect/EXPERTconnect/UI/Sections/Navigation/ECSDynamicViewController.m
@@ -102,7 +102,7 @@ static NSString *const ECSQuestionCellId = @"ECSQuestionCellId";
      one nav item exists, then don't display the navigation UI and instead launch
      the Action. */
     if (self.navigationContext.sections.count == 1) {
-        int actionCount = 0;
+        NSUInteger actionCount = 0;
         ECSNavigationSection *navSection = self.navigationContext.sections[0];
         ECSActionType *actiontype = nil;
         switch (navSection.sectionType) {
